Adds timeout to MdbMaster::GetResponse().
Timeout is set to 5ms after sending a command. If for some reason the
GetResponse method is called more than 5ms after the command is sent,
but data is available, the command should still be received (because the
method checks if data is available before checking the timeout). On the
flip side, if some method decides to call GetResponse when no command
has been sent, it should just fail pretty quickly, returning a NAK,
because the timeout will have expired (if that's even meaningful).