Fixes GetFeatureLevel and adds getters.
Fixes the implementation of the GetFeatureLevel. It had a bug with
retrieving the correct number of characters. I assumed that it would
always return 27 bytes, but the number of bytes returned actually
depends on the number of bill types supported.

Also adds getter methods for each property of the connected bill
acceptor, as well as a ToString method that returns all of these
properties as a formatted string.